--- client/src/reducers/userreducer.js
@@ -1,20 +1,0 @@
-//function containing object for restaurant state data
-const defaultState = {  
-    id: 1,
-    name: 'Anonymous',
-    location: 'Nowhere, CA',
-    bio: 'blah, blah',
-}
-
-
-export default  function (state= defaultState, action){
-    switch(action.type) {
-        case "EDIT_BIO":
-          return Object.assign({}, state, { bio: action.payload } )
-          break;
-        default:
-          return state;
-    
-      }
-      return state;
-     }
