+++ client/src/store/reducers/reviewsreducer.js
@@ -1,30 +1,38 @@
-//function containing object for restaurant state data
 //dummy reviews data
-export default function () {
-    return [
-        {
-            id: 1,
-            userid: 1,
-            score: 5,
-            restaurantid: 2,
-            text: "Bucky is a React developer and YouTuber"
-           
-        },
-        {
-            id: 2,
-            userid: 2,
-            score: 5,
-            restaurantid: 2,
-            text: "Bucky is a React developer and YouTuber"
-        },
-        {
-            id: 3,
-            userid: 3,
-            score: 5,
-            restaurantid: 2,
-            text: "Bucky is a React developer and YouTuber"
-        }
-    ]
+const dumreviews =  [
+    {
+        id: 1,
+        userid: 1,
+        score: 5,
+        restaurantid: 2,
+        text: "Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa."
+       
+    },
+    {
+        id: 2,
+        userid: 2,
+        score: 5,
+        restaurantid: 2,
+        text: "Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa."
+    },
+    {
+        id: 3,
+        userid: 1,
+        score: 5,
+        restaurantid: 2,
+        text: "Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa."
+    }
+];
+
+export default function (state=dumreviews, action) {
+    switch(action.type) {
+        case "INIT_REVIEWS":
+            console.log('payload', action.payload);
+          return action.payload
+          break;
+        default:
+          return state;
+    
+      }
 }
 
-//export default restaurantsreducer;
